+++ src/wrapper/jnienv.rs
@@ -8,6 +8,7 @@ use std::{
 
 use log::warn;
 
+use crate::signature::ReturnType;
 use crate::{
     descriptors::Desc,
     errors::*,
@@ -765,8 +766,8 @@ impl<'a> JNIEnv<'a> {
         &self,
         class: T,
         method_id: U,
-        ret: JavaType,
-        args: &[JValue],
+        ret: ReturnType,
+        args: &[jvalue],
     ) -> Result<JValue<'a>>
     where
         T: Desc<'a, JClass<'c>>,
@@ -777,12 +778,11 @@ impl<'a> JNIEnv<'a> {
         let method_id = method_id.lookup(self)?.into_inner();
 
         let class = class.into_inner();
-        let args: Vec<jvalue> = args.iter().map(|v| v.to_jni()).collect();
         let jni_args = args.as_ptr();
 
         // TODO clean this up
         Ok(match ret {
-            JavaType::Object(_) | JavaType::Array(_) => {
+            ReturnType::Object | ReturnType::Array => {
                 let obj: JObject = jni_non_void_call!(
                     self.internal,
                     CallStaticObjectMethodA,
@@ -793,9 +793,7 @@ impl<'a> JNIEnv<'a> {
                 .into();
                 obj.into()
             }
-            // JavaType::Object
-            JavaType::Method(_) => unimplemented!(),
-            JavaType::Primitive(p) => match p {
+            ReturnType::Primitive(p) => match p {
                 Primitive::Boolean => jni_non_void_call!(
                     self.internal,
                     CallStaticBooleanMethodA,
@@ -884,8 +882,8 @@ impl<'a> JNIEnv<'a> {
         &self,
         obj: O,
         method_id: T,
-        ret: JavaType,
-        args: &[JValue],
+        ret: ReturnType,
+        args: &[jvalue],
     ) -> Result<JValue<'a>>
     where
         O: Into<JObject<'a>>,
@@ -895,20 +893,17 @@ impl<'a> JNIEnv<'a> {
 
         let obj = obj.into().into_inner();
 
-        let args: Vec<jvalue> = args.iter().map(|v| v.to_jni()).collect();
         let jni_args = args.as_ptr();
 
         // TODO clean this up
         Ok(match ret {
-            JavaType::Object(_) | JavaType::Array(_) => {
+            ReturnType::Object | ReturnType::Array => {
                 let obj: JObject =
                     jni_non_void_call!(self.internal, CallObjectMethodA, obj, method_id, jni_args)
                         .into();
                 obj.into()
             }
-            // JavaType::Object
-            JavaType::Method(_) => unimplemented!(),
-            JavaType::Primitive(p) => match p {
+            ReturnType::Primitive(p) => match p {
                 Primitive::Boolean => {
                     jni_non_void_call!(self.internal, CallBooleanMethodA, obj, method_id, jni_args)
                         .into()
@@ -984,7 +979,8 @@ impl<'a> JNIEnv<'a> {
 
         let class = self.auto_local(self.get_object_class(obj)?);
 
-        self.call_method_unchecked(obj, (&class, name, sig), parsed.ret, args)
+        let args: Vec<jvalue> = args.iter().map(|v| v.to_jni()).collect();
+        self.call_method_unchecked(obj, (&class, name, sig), parsed.ret, &args)
     }
 
     /// Calls a static method safely. This comes with a number of
@@ -1019,7 +1015,8 @@ impl<'a> JNIEnv<'a> {
         // and we'll need that for the next call.
         let class = class.lookup(self)?;
 
-        self.call_static_method_unchecked(class, (class, name, sig), parsed.ret, args)
+        let args: Vec<jvalue> = args.iter().map(|v| v.to_jni()).collect();
+        self.call_static_method_unchecked(class, (class, name, sig), parsed.ret, &args)
     }
 
     /// Create a new object using a constructor. This is done safely using
@@ -1041,7 +1038,7 @@ impl<'a> JNIEnv<'a> {
             return Err(Error::InvalidArgList(parsed));
         }
 
-        if parsed.ret != JavaType::Primitive(Primitive::Void) {
+        if parsed.ret != ReturnType::Primitive(Primitive::Void) {
             return Err(Error::InvalidCtorReturn);
         }
 
@@ -1682,7 +1679,7 @@ impl<'a> JNIEnv<'a> {
         &self,
         obj: O,
         field: T,
-        ty: JavaType,
+        ty: ReturnType,
     ) -> Result<JValue<'a>>
     where
         O: Into<JObject<'a>>,
@@ -1696,14 +1693,12 @@ impl<'a> JNIEnv<'a> {
 
         // TODO clean this up
         Ok(match ty {
-            JavaType::Object(_) | JavaType::Array(_) => {
+            ReturnType::Object | ReturnType::Array => {
                 let obj: JObject =
                     jni_non_void_call!(self.internal, GetObjectField, obj, field).into();
                 obj.into()
             }
-            // JavaType::Object
-            JavaType::Method(_) => unimplemented!(),
-            JavaType::Primitive(p) => match p {
+            ReturnType::Primitive(p) => match p {
                 Primitive::Boolean => {
                     jni_unchecked!(self.internal, GetBooleanField, obj, field).into()
                 }
@@ -1784,7 +1779,7 @@ impl<'a> JNIEnv<'a> {
         let obj = obj.into();
         let class = self.auto_local(self.get_object_class(obj)?);
 
-        let parsed = JavaType::from_str(ty.as_ref())?;
+        let parsed = ReturnType::from_str(ty.as_ref())?;
 
         let field_id: JFieldID = (&class, name, ty).lookup(self)?;
 
@@ -1957,7 +1952,7 @@ impl<'a> JNIEnv<'a> {
         // Check to see if we've already set this value. If it's not null, that
         // means that we're going to leak memory if it gets overwritten.
         let field_ptr = self
-            .get_field_unchecked(obj, field_id, JavaType::Primitive(Primitive::Long))?
+            .get_field_unchecked(obj, field_id, ReturnType::Primitive(Primitive::Long))?
             .j()? as *mut Mutex<T>;
         if !field_ptr.is_null() {
             return Err(Error::FieldAlreadySet(field.as_ref().to_owned()));
@@ -2012,7 +2007,7 @@ impl<'a> JNIEnv<'a> {
             let guard = self.lock_obj(obj)?;
 
             let ptr = self
-                .get_field_unchecked(obj, field_id, JavaType::Primitive(Primitive::Long))?
+                .get_field_unchecked(obj, field_id, ReturnType::Primitive(Primitive::Long))?
                 .j()? as *mut Mutex<T>;
 
             non_null!(ptr, "rust value from Java");