+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Default` trait implemented for `JObject`, `JString`, `JClass`, and `JByteBuffer` (#199)
 - `JNIEnv#new_direct_byte_buffer_raw` function that allows creating a `JByteBuffer` from a pointer and size (#351)
 - `Debug` trait implemented for `JavaVM`, `GlobalRef`, `GlobalRefGuard`, `JStaticMethodID` and `ReleaseMode`
+- `ReturnType` for specifying object return types without a String allocation. (#329)
 
 ### Changed
 - The `release_string_utf_chars` function has been marked as unsafe. (#334)
@@ -28,6 +29,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   reference from a global reference. (#301 / #319)
 - `JMethodID` implements `Send` + `Sync` and no longer has a lifetime parameter, making method
   IDs cacheable (with a documented 'Safety' note about ensuring they remain valid). ([#346](https://github.com/jni-rs/jni-rs/pull/346))
+- The `call_*_method_unchecked` functions now take `jni:sys::jvalue` arguments to avoid allocating
+  a `Vec` on each call to map + collect `JValue`s as `sys:jvalue`s (#329)
 
 ## [0.19.0] — 2021-01-24